requirements.txt
flask
python-dotenv
openai
redis